requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astor==0.8.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
Babel==2.15.0
bce-python-sdk==0.9.19
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.112.0
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.53.1
frozenlist==1.4.1
future==1.0.0
greenlet==3.0.3
grpcio==1.65.4
grpcio-tools==1.65.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
imageio==2.34.2
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.2.12
langchain-community==0.2.10
langchain-core==0.2.27
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langsmith==0.1.96
lazy_loader==0.4
lmdb==1.5.1
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
more-itertools==10.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
openai==1.38.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.6
packaging==24.1
paddleocr==2.8.1
pandas==2.2.2
pdf2docx==0.5.8
pillow==10.4.0
portalocker==2.10.1
premailer==3.10.0
protobuf==5.27.3
psutil==6.0.0
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
# PyMuPDF==1.20.2
PyMySQL==1.0.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.10.1
rapidfuzz==3.9.5
rarfile==4.2
regex==2024.7.24
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.53
starlette==0.37.2
tenacity==8.5.0
termcolor==2.4.0
tifffile==2024.7.24
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visualdl==2.5.3
Werkzeug==3.0.3
yarl==1.9.4
paddlepaddle==2.6.2
pypdf==5.0.0
docx2txt==0.8